requirements.txt
# CrewAI and Core Dependencies
crewai[tools]==0.175.0
crewai-tools==0.12.1
pydantic==2.9.2

# LLM and API Integration
litellm==1.52.3
mistralai==1.2.0

# Web Scraping and Browser Automation
requests==2.32.3
beautifulsoup4==4.12.2
selenium==4.27.1
webdriver-manager==4.0.2
playwright==1.48.0
browser-use==0.1.17

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Environment and Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Additional Tools
pyyaml==6.0.2
typing-extensions==4.12.2
